+++ src/fiqci/ems/transpiler_passes/pauli_twirl.py
@@ -95,7 +95,7 @@ def get_twirled_circuits(
 	twirled_circuits = []
 
 	if backend is not None:
-		basis_gates = list({i[0].name for i in backend.target.instructions})
+		basis_gates = list({i[0].name for i in backend.target.instructions if isinstance(i[0].name, str)})
 		pm = PassManager(
 			[
 				PauliTwirl(gates_to_twirl=gates_to_twirl),